Create Python Program for Tic-Tac-Toe
The task is to create a simple Tic-Tac-Toe game using a 2D array to represent the game board. The game will feature two players, X and O, who will take turns making moves. The program will also check for winning conditions and determine if the game ends in a draw.